Delayed Recoils - Fixed Pipe Trigger Bug - Forced Standstill during stab
Tongue only retracts when your animation reaches the stabbing part.
You can now poison the pipes properly during the healing phase, before
if you were adjacent to a pipe prior his switch to healing, half the
conditions for enabling stabbing were not met, so this has now been
adjusted to work in every case.
Finally, when stabbing, the player is now stationary for the duration of
the animatino, so you have to finish it before being allowed to move